Reorganize dropdown editors into Editors/dropdown/
Consolidate selector components and existing boolean/vocabulary editor
folders under Editors/dropdown/<domain>/. Rename *Selector files to
*Dropdown to match underlying primitive.
