+++ src/main/cliapp/src/components/Editors/dropdown/vocabulary/ControlledVocabularyMultiSelectDropdown.jsx
