+++ src/main/cliapp/src/containers/allelesPage/AllelesTable.jsx
@@ -32,7 +32,7 @@ import { DialogTriggerEditor } from '../../components/Editors/DialogTriggerEdito
 import { TaxonTableEditor } from '../../components/Editors/taxon/TaxonTableEditor';
 import { InCollectionTableEditor } from '../../components/Editors/inCollection/InCollectionTableEditor';
 import { ReferencesTableEditor } from '../../components/Editors/references/ReferencesTableEditor';
-import { BooleanTableEditor } from '../../components/Editors/boolean/BooleanTableEditor';
+import { BooleanTableEditor } from '../../components/Editors/dropdown/boolean/BooleanTableEditor';
 
 import { TruncatedReferencesTemplate } from '../../components/Templates/reference/TruncatedReferencesTemplate';
 import { IdTemplate } from '../../components/Templates/IdTemplate';
