+++ src/main/cliapp/src/containers/controlledVocabularyPage/ControlledVocabularyTable.jsx
@@ -6,8 +6,8 @@ import { Toast } from 'primereact/toast';
 import { useControlledVocabularyService } from '../../service/useControlledVocabularyService';
 import { VocabularyService } from '../../service/VocabularyService';
 import { InputTextTableEditor } from '../../components/Editors/text/InputTextTableEditor';
-import { ControlledVocabularyTableEditor } from '../../components/Editors/controlledVocabulary/ControlledVocabularyTableEditor';
-import { BooleanTableEditor } from '../../components/Editors/boolean/BooleanTableEditor';
+import { ControlledVocabularyTableEditor } from '../../components/Editors/dropdown/vocabulary/ControlledVocabularyTableEditor';
+import { BooleanTableEditor } from '../../components/Editors/dropdown/boolean/BooleanTableEditor';
 import { NewTermForm } from '../../containers/controlledVocabularyPage/NewTermForm';
 import { NewVocabularyForm } from '../../containers/controlledVocabularyPage/NewVocabularyForm';
 import { Button } from 'primereact/button';
