+++ src/main/cliapp/src/containers/diseaseAnnotationsPage/DiseaseAnnotationsTable.jsx
@@ -24,11 +24,11 @@ import { CountDialogTemplate } from '../../components/Templates/dialog/CountDial
 import { TextDialogTemplate } from '../../components/Templates/dialog/TextDialogTemplate';
 import { StringTemplate } from '../../components/Templates/StringTemplate';
 
-import { NotEditor } from '../../components/Editors/NotEditor';
-import { ControlledVocabularyTableEditor } from '../../components/Editors/controlledVocabulary/ControlledVocabularyTableEditor';
-import { BooleanTableEditor } from '../../components/Editors/boolean/BooleanTableEditor';
+import { NotEditor } from '../../components/Editors/dropdown/not/NotEditor';
+import { ControlledVocabularyTableEditor } from '../../components/Editors/dropdown/vocabulary/ControlledVocabularyTableEditor';
+import { BooleanTableEditor } from '../../components/Editors/dropdown/boolean/BooleanTableEditor';
 import { DialogTriggerEditor } from '../../components/Editors/DialogTriggerEditor';
-import { ConditionRelationHandleTableEditor } from '../../components/Editors/ConditionRelationHandleTableEditor';
+import { ConditionHandleTableEditor } from '../../components/Editors/dropdown/conditionHandle/ConditionHandleTableEditor';
 import { BiologicalEntityTableEditor } from '../../components/Editors/biologicalEntity/BiologicalEntityTableEditor';
 import { DiseaseTableEditor } from '../../components/Editors/ontology/DiseaseTableEditor';
 import { EvidenceCodesTableEditor } from '../../components/Editors/ontology/EvidenceCodesTableEditor';
@@ -40,7 +40,7 @@ import { DiseaseGeneticModifierAllelesTableEditor } from '../../components/Edito
 import { SgdStrainBackgroundTableEditor } from '../../components/Editors/agm/SgdStrainBackgroundTableEditor';
 import { DiseaseGeneticModifierAgmsTableEditor } from '../../components/Editors/agm/DiseaseGeneticModifierAgmsTableEditor';
 import { SingleReferenceTableEditor } from '../../components/Editors/references/SingleReferenceTableEditor';
-import { ControlledVocabularyMultiSelectTableEditor } from '../../components/Editors/controlledVocabulary/ControlledVocabularyMultiSelectTableEditor';
+import { ControlledVocabularyMultiSelectTableEditor } from '../../components/Editors/dropdown/vocabulary/ControlledVocabularyMultiSelectTableEditor';
 
 import { useControlledVocabularyService } from '../../service/useControlledVocabularyService';
 import { useVocabularyTermSetService } from '../../service/useVocabularyTermSetService';
@@ -379,7 +379,7 @@ export const DiseaseAnnotationsTable = () => {
 				sortable: true,
 				filterConfig: FILTER_CONFIGS.daConditionRelationsHandleFilterConfig,
 				editor: (editorOptions) => (
-					<ConditionRelationHandleTableEditor editorOptions={editorOptions} errorMessagesRef={errorMessagesRef} />
+					<ConditionHandleTableEditor editorOptions={editorOptions} errorMessagesRef={errorMessagesRef} />
 				),
 			},
 			{
