+++ src/main/cliapp/src/containers/diseaseAnnotationsPage/NewAnnotationForm.jsx
@@ -6,7 +6,7 @@ import { Toast } from 'primereact/toast';
 import { MultiSelect } from 'primereact/multiselect';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { FormErrorMessageComponent } from '../../components/Error/FormErrorMessageComponent';
-import { NotEditor } from '../../components/Editors/NotEditor';
+import { NotEditor } from '../../components/Editors/dropdown/not/NotEditor';
 import { classNames } from 'primereact/utils';
 import { DiseaseAnnotationService } from '../../service/DiseaseAnnotationService';
 import { Splitter, SplitterPanel } from 'primereact/splitter';
@@ -15,10 +15,10 @@ import { SubjectAutocompleteTemplate } from '../../components/Autocomplete/Subje
 import { EvidenceAutocompleteTemplate } from '../../components/Autocomplete/EvidenceAutocompleteTemplate';
 import { RelatedNotesForm } from './RelatedNotesForm';
 import { ConditionRelationsForm } from './ConditionRelationsForm';
-import { ConditionRelationHandleFormDropdown } from '../../components/ConditionRelationHandleFormSelector';
-import { ControlledVocabularyFormDropdown } from '../../components/ControlledVocabularyFormSelector';
+import { ConditionHandleFormDropdown } from '../../components/Editors/dropdown/conditionHandle/ConditionHandleFormDropdown';
+import { ControlledVocabularyFormDropdown } from '../../components/Editors/dropdown/vocabulary/ControlledVocabularyFormDropdown';
 import { useControlledVocabularyService } from '../../service/useControlledVocabularyService';
-import { ControlledVocabularyFormMultiSelectDropdown } from '../../components/ControlledVocabularyFormMultiSelector';
+import { ControlledVocabularyFormMultiSelectDropdown } from '../../components/Editors/dropdown/vocabulary/ControlledVocabularyFormMultiSelectDropdown';
 import { AutocompleteFormEditor } from '../../components/Autocomplete/AutocompleteFormEditor';
 import {
 	autocompleteSearch,
@@ -949,7 +949,7 @@ export const NewAnnotationForm = ({
 										<label htmlFor="experiments">{required}Experiments</label>
 									</div>
 									<div className={widgetColumnSize}>
-										<ConditionRelationHandleFormDropdown
+										<ConditionHandleFormDropdown
 											name="experiments"
 											customRef={experimentsRef}
 											editorChange={onDropdownExperimentsFieldChange}
